+++ src/geo_activity_playground/features/strava/api_importer.py
@@ -13,6 +13,7 @@ from tqdm import tqdm
 
 from ...core.config import ConfigAccessor
 from ...core.datamodel import DB, Activity, get_or_make_equipment, get_or_make_kind
+from ...core.duplicate_matching import check_for_duplicate
 from ...core.enrichment import update_and_commit
 from ...core.import_exclusion import is_excluded
 from ...core.paths import (
@@ -264,6 +265,7 @@ def try_import_strava(
                 update_and_commit(
                     activity, time_series, config_accessor.activity_import()
                 )
+                check_for_duplicate(activity, config_accessor.activity_import())
                 logger.info(f"Added activity '{activity.name}' from Strava.")
 
             if strava_begin is None and strava_end is None: